fitness-coach-bot/requirements.txt
python-telegram-bot[job-scheduling]==21.10
anthropic>=0.42.0
notion-client>=2.2.0
python-dotenv>=1.0.0
pytz>=2024.1
